This commit was generated by cvs2svn to compensate for changes in r11861, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@11862 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/raster/raster-canvas.c
@@ -81,8 +81,8 @@ inline static void refresh_canvas(raster_t *raster)
     yy += viewport->y_offset;
 
     video_canvas_refresh(raster->canvas, x, y, xx, yy,
-        MIN(w, raster->canvas->draw_buffer->canvas_width - xx),
-        MIN(h, raster->canvas->draw_buffer->canvas_height - yy));
+        MIN(w, (int)(raster->canvas->draw_buffer->canvas_width - xx)),
+        MIN(h, (int)(raster->canvas->draw_buffer->canvas_height - yy)));
     update_area->is_null = 1;
 }
